refactor: as typescript
